+++ types/sidemenu.ts
@@ -1,0 +1,18 @@
+export interface SideMenuListItemProps {
+  id: number
+  color: string
+  title: string
+  createdByMe: boolean
+}
+
+export interface DashBoardProps {
+  id: number
+  color: string
+  title: string
+  createdByMe: boolean
+}
+
+export interface SideMenuListProps {
+  // eslint-disable-next-line react/require-default-props
+  data: DashBoardProps[] | undefined
+}
